+++ blas/CMakeLists.txt
@@ -7,4 +7,9 @@ add_library(blas SHARED
             ${SOURCE_FILES_PROJECT_GLOBALS}
             blas.h
             common_blas.h
-            Tensor.h Tensor.cpp TensorMath.h TensorMath.cpp TensorCreation.h TensorCreation.cpp)
+            Slice.h
+            Tensor.h TensorView.h TensorSliced.h TensorTransposed.h
+            all_tensors.h
+            implementation.cpp
+            TensorMath.h TensorMath.cpp 
+            TensorCreation.h TensorCreation.cpp)
