+++ blas/Tensor.h
@@ -2,8 +2,10 @@
 // Created by LevZ on 7/8/2020.
 //
 
-#ifndef TARGETPRACTICE_TENSOR_H
-#define TARGETPRACTICE_TENSOR_H
+#ifndef BLAS_TENSOR_H_
+#define BLAS_TENSOR_H_
+
+ 
 
 #include <sstream>
 #include <stdexcept>
@@ -12,20 +14,13 @@
 #include <vector>
 
 #include "common_blas.h"
+#include "Slice.h"
 
 using namespace common_math;
 
-std::string shape2str(const shape_t& shape);
-std::string shape2str(const std::vector<long>& shape);
-size_t shape2size(const shape_t& shape);
 
 // Math operations on tensors, like numpy.
 namespace blas {
-using std::initializer_list;
-using std::ostream;
-using std::string;
-using std::tuple;
-using std::vector;
 
 template <typename T>
 class Tensor;
@@ -36,70 +31,12 @@ class TensorView;
 template <typename T>
 class TensorSliced;
 
+// template <typename T>
+// class TensorSparse;
+
 template <typename T>
-class TensorSparse;
+class TensorTransposed;
 
-shape_t shape2strides(const shape_t& shape);
-
-/**
- * Translates a dims-format index to a tuple of true index and the number of
- * elements associated with it. Checks for out-of-range errors.
- * @param idx: Unraveled index
- * @param shape: Shape of tensor
- * @param size : the total size of the tensor. Optional, if (-1) the size will
- * be determined from the shape.
- * @return pair of (true_index, num_elements).
- */
-std::pair<size_t, size_t> ravel_index_checked(const index_t& idx,
-                                              const shape_t& shape,
-                                              int size = -1);
-
-/**
- * @brief Translates a dims-format index to a tuple of true index and the number
- * of elements associated with it. Doesn't check for errors and only returns the
- * true index.
- * @param idx : see ravel_index_checked
- * @param shape : see ravel_index_checked
- * @param size : see ravel_index_checked
- * @return true_idx
- */
-size_t ravel_index(const index_t& idx, const shape_t& shape, int size = -1);
-
-/**
- * The invers operation of ravel index - translates a true index into
- * dims-format index.
- * @param true_idx : true array index.
- * @param shape : see ravel_index_checked
- * @param size : see ravel_index_checked
- * @return unraveled index for the current tensor
- */
-index_t unravel_index(size_t true_idx, const shape_t& shape, int size = -1);
-
-class shape_mismatch : public std::out_of_range {
-   public:
-    shape_mismatch(const shape_t& s1, const shape_t& s2,
-                   const std::string& action = "")
-        : std::out_of_range(
-              "Shape Mismatch: " + shape2str(s1) + ", " + shape2str(s2) +
-              (action.empty() ? "" : " for action {" + action + "}") + ".") {}
-
-    shape_mismatch(const vector<long>& s1, const shape_t& s2,
-                   const std::string& action = "")
-        : std::out_of_range(
-              "Shape Mismatch: " + shape2str(s1) + ", " + shape2str(s2) +
-              (action.empty() ? "" : " for action {" + action + "}") + ".") {}
-};
-
-class broadcast_failure : public std::out_of_range {
-   public:
-    using std::out_of_range::out_of_range;
-    broadcast_failure(const shape_t& s1, const shape_t& s2,
-                      const std::string& action = "")
-        : std::out_of_range(
-              "Cannot Broadcast shapes: " + shape2str(s1) + ", " +
-              shape2str(s2) +
-              (action.empty() ? "" : " for action {" + action + "}") + ".") {}
-};
 
 /**
  * Fill a tensor with a single value inplace, generically.
@@ -210,217 +147,6 @@ Tensor1& copy_(Tensor1& dst, const Tensor2& src);
         return *this;                            \
     }                                            \
     MACRO_INTERACTABLE_TENSORTYPES(DEF_ASSIGNMENT_TEMPLATE, Tensor1, T)
-
-class Slice {
-   private:
-    class const_iterator {
-        using index_t = long;
-        index_t pos;
-        index_t stride;
-
-       public:
-        using difference_type = long;  // unused
-        using value_type = index_t;
-        using pointer = const index_t*;  // unused
-        using reference = const index_t&;
-        using iterator_category = std::input_iterator_tag;
-
-        const_iterator(index_t pos, index_t stride)
-            : pos(pos), stride(stride) {}
-
-        inline const_iterator& operator++() {
-            pos += stride;
-            return *this;
-        }
-
-        inline const const_iterator operator++(int) {
-            const_iterator temp(*this);
-            ++*this;
-            return temp;
-        }
-
-        inline const_iterator& operator+=(difference_type x) {
-            pos += x * stride;
-            return *this;
-        }
-
-        inline const_iterator operator+(difference_type x) {
-            const_iterator temp(*this);
-            return temp += x;
-        }
-
-        inline bool operator==(const const_iterator& other) const {
-            return pos == other.pos;
-        }
-
-        inline bool operator!=(const const_iterator& other) const {
-            return !(*this == other);
-        }
-
-        inline reference operator*() const { return pos; }
-    };
-
-    /* why isn't abs constexpr -_- */
-    static inline constexpr size_t abs(long x) { return x < 0 ? -x : x; }
-    size_t size_;
-    static inline constexpr size_t get_size(long b, long e, long stride) {
-        return (abs(e - b) - 1) / abs(stride) + 1;
-    }
-
-   public:
-    long b, e, stride;
-
-    Slice() : Slice(0, 1, 1) {}
-    Slice(long b, long e, long stride = 1);
-
-    explicit Slice(const tuple<long, long, long>& tup)
-        : Slice(std::get<0>(tup), std::get<1>(tup), std::get<2>(tup)) {}
-
-    Slice(initializer_list<long> lst);
-
-    constexpr void check_stride() const;
-
-    inline size_t size() const { return size_; }
-
-    inline const_iterator begin() const { return const_iterator(b, stride); }
-
-    inline const_iterator end() const { return const_iterator(e, stride); }
-
-    inline void update() { size_ = get_size(b, e, stride); }
-
-    string to_str() const {
-        using std::to_string;
-        return to_string(b) + ":" + to_string(e) +
-               (stride == 1 ? "" : (":" + to_string(stride)));
-    }
-
-    Slice subslice(const Slice& relative_slice) const;
-};
-
-/**
- * An object that represents an index iterator over a Slice of a tensor.
- */
-class SliceGroup {
-   public:
-    vector<Slice> slices;
-    SliceGroup() = default;
-    SliceGroup(size_t dims) : slices(dims) {}
-    explicit SliceGroup(const vector<tuple<int, int, int>>& slices);
-    explicit SliceGroup(const vector<Slice>& slices) : slices(slices) {}
-    SliceGroup(initializer_list<initializer_list<long>> lst);
-
-    static inline SliceGroup cover_shape(const shape_t& shape) {
-        SliceGroup ret;
-        ret.slices.reserve(shape.size());
-        for (size_t x : shape) ret.slices.emplace_back(0, x);
-        return ret;
-    }
-    static inline SliceGroup cover_index(const index_t& index) {
-        SliceGroup ret;
-        ret.slices.reserve(index.size());
-        for (long x : index) {
-            ret.slices.emplace_back(x, x + 1);
-        }
-        return ret;
-    }
-
-    inline SliceGroup fill_to_shape(const shape_t& cover_shape) const {
-        SliceGroup ret(*this);
-        return ret.fill_to_shape_(cover_shape);
-    }
-    inline SliceGroup& fill_to_shape_(const shape_t& cover_shape) {
-        size_t old_size = slices.size();
-        slices.reserve(cover_shape.size());
-        for (size_t i = old_size; i < cover_shape.size(); ++i)
-            slices.emplace_back(0, cover_shape[i]);
-
-        return *this;
-    }
-
-    class const_iterator {
-        vector<Slice> slices;
-        index_t pos;
-        size_t elems_passed;
-
-       public:
-        using difference_type = long;  // unused
-        using value_type = index_t;
-        using pointer = const index_t*;  // unused
-        using reference = const index_t&;
-        using iterator_category = std::input_iterator_tag;
-
-        const_iterator(index_t pos, vector<Slice> slices,
-                       size_t elems_passed = 0);
-
-        const_iterator& operator++();
-        inline const_iterator operator++(int) {
-            const_iterator temp(*this);
-            ++*this;
-            return temp;
-        }
-
-        inline bool operator==(const const_iterator& other) const {
-            return elems_passed == other.elems_passed;
-        }
-
-        inline bool operator!=(const const_iterator& other) const {
-            return !(*this == other);
-        }
-
-        inline reference operator*() { return pos; }
-    };
-    inline const_iterator begin() const {
-        return const_iterator(get_init_pos(), slices, 0);
-    }
-    inline const_iterator end() const {
-        return const_iterator(get_init_pos(), slices, size());
-    }
-
-    inline index_t translate_true_idx(size_t true_index) const {
-        index_t ret(slices.size());
-        for (int i = ret.size() - 1; i >= 0; --i) {
-            auto s = slices[i];
-            if (s.size() != 1) {
-                ret[i] = s.b + (true_index % s.size()) * s.stride;
-                true_index /= s.size();
-            } else
-                ret[i] = s.b;
-        }
-        return ret;
-    }
-    inline size_t translate_vector_idx(index_t vec_index) const {
-        size_t true_index = 0;
-        size_t tot_stride = 1;
-        for (int i = slices.size() - 1; i >= 0; --i) {
-            auto s = slices[i];
-            if (s.size()) {
-                long p = vec_index[i];
-                size_t pos = (p - s.b) / s.stride;
-                true_index += pos * tot_stride;
-                tot_stride *= s.size();
-            }
-        }
-        return true_index;
-    }
-    inline shape_t shape() const {
-        shape_t ret(slices.size());
-        for (int i = 0; i < ret.size(); ++i) ret[i] = slices[i].size();
-        return ret;
-    }
-    SliceGroup subslice(const SliceGroup& relative_slice) const;
-    size_t size() const;
-    string to_str() const;
-
-    inline void update() {}
-
-   private:
-    inline vector<long> get_init_pos() const {
-        vector<long> res(slices.size());
-        for (int i = 0; i < res.size(); ++i) res[i] = slices[i].b;
-        return res;
-    }
-};
-
 template <typename T>
 class Tensor {
    public:
@@ -732,287 +458,9 @@ class Tensor {
 
 template <typename T>
 size_t Tensor<T>::print_precision = 0;
-/**
- * A view of a tensor - doesn't copy data from original tensor.
- * @tparam T : stored data type.
- */
-template <typename T>
-class TensorView : public Tensor<T> {
-    friend class Tensor<T>;
-
-    TensorView(T* data, const std::vector<size_t>&);
-
-    explicit TensorView(Tensor<T> t);
-
-   public:
-    ~TensorView() override = default;  // Doesn't delete the data.
-    using eiterator = T*;
-    using ceiterator = const T*;
-
-    static T& get(TensorView<T>& tv, size_t true_idx) {
-        return tv.data[true_idx];
-    }
-    static T get(const TensorView<T>& tv, size_t true_idx) {
-        return tv.data[true_idx];
-    }
-
-    inline Tensor<T> contiguous() override {
-        Tensor<T> t(this->shape);
-        t.copy_(*this);
-        return t;
-    }
-
-    // We don't need this here because it operates the same exact way as for
-    // Tensor.
-    // // Declares all interactions between the different tensor types.
-    // MACRO_INTERACTABLE_TENSORTYPES(DECL_INTERACTIVE_ACTIONS_TENSOR_OVERRIDE,
-    // TensorView, T)
-
-    DECL_INTERACTIVE_ACTION_TENSOR_UNIQUE_OVERRIDE(TensorView)
-    DEF_COPY_FILL_TEMPLATES(TensorView, T)
-};
-
-/**
- * A slice of the tensor - doesn't copy data from the original tensor.
- * @tparam T
- */
-template <typename T>
-class TensorSliced : public Tensor<T> {
-   private:
-    friend class Tensor<T>;
-    shape_t underlying_tensor_shape;
-    const size_t underlying_tensor_size;
-
-    TensorSliced(T* data, const shape_t& shape, const SliceGroup& slice_group);
-
-    TensorSliced(const Tensor<T>& t, const SliceGroup& slice_group);
-
-   public:
-    ~TensorSliced() override = default;
-    TensorSliced(const TensorSliced& other);
-    TensorSliced(TensorSliced&& other) noexcept = default;
-
-    DECL_INTERACTIVE_ACTION_TENSOR_UNIQUE_OVERRIDE(TensorSliced)
-    DEF_COPY_FILL_TEMPLATES(TensorSliced, T)
-
-    template <typename T_>
-    struct eliterator {
-        T_* data;
-        SliceGroup::const_iterator sg_iterator;
-        const shape_t shape;
-        const size_t size;
-
-        using difference_type = ptrdiff_t;
-        using value_type = T_;
-        using reference = T_&;
-        using pointer = T_*;
-        using iterator_category = std::forward_iterator_tag;
-
-        inline reference operator*() {
-            const auto& idx = *sg_iterator;
-            auto true_idx = ravel_index(idx, shape, size);
-            return data[true_idx];
-        }
-
-        inline eliterator& operator++() {
-            ++sg_iterator;
-            return *this;
-        }
-
-        inline eliterator operator++(int) {
-            eliterator temp(*this);
-            ++(*this);
-            return temp;
-        }
-
-        inline bool operator==(const eliterator& other) {
-            return sg_iterator == other.sg_iterator;
-        }
-
-        inline bool operator!=(const eliterator& other) {
-            return !(*this == other);
-        }
-    };
-    using eiterator = eliterator<T>;
-    using ceiterator = eliterator<const T>;
-    static inline eiterator elem_begin(TensorSliced& ts);
-    static inline eiterator elem_end(TensorSliced& ts);
-    static inline ceiterator const_elem_begin(const TensorSliced& ts);
-    static inline ceiterator const_elem_end(const TensorSliced& ts);
-
-    /* We mark these as forbidden because we cannot create TensorView out of
-     * TensorSlice.*/
-    MARK_FORBIDDEN(TensorView<T> unchecked_subscript(long idx) const override)
-    MARK_FORBIDDEN(TensorView<T> unchecked_subscript(const index_t& index)
-                       const override)
-    TensorSliced<T> unchecked_slice(const Slice& slice) const override;
-    TensorSliced<T> unchecked_slice_group(
-        const SliceGroup& slice_group) const override;
-    MARK_FORBIDDEN(TensorView<T> view(const vector<long>& new_shape) override)
-
-#define DECL_TENSOR_REDUCE_OVERRIDE(TensorOut)                               \
-    void reduce(const binary_op<T>& op, TensorOut<T>& out) const override;   \
-    void reduce(const binary_op<T>& op, int dim, TensorOut<T>& out)          \
-        const override;                                                      \
-    void reduce(const binary_op<T>& op, vector<int> dims, TensorOut<T>& out) \
-        const override;
-
-    DECL_TENSOR_REDUCE_OVERRIDE(Tensor)
-    DECL_TENSOR_REDUCE_OVERRIDE(TensorView)
-    DECL_TENSOR_REDUCE_OVERRIDE(TensorSliced)
-
-    Tensor<T> reshape(const vector<long>& new_shape) const override;
-
-    Tensor<T> contiguous() override;
-
-    SliceGroup slice_group;
-
-   protected:
-    ostream& print_to_os(ostream& os, bool rec_start) const override;
-
-   public:
-    static T& get(TensorSliced<T>& ts, size_t true_idx);
-    static T get(const TensorSliced<T>& ts, size_t true_idx);
-
-    MACRO_INTERACTABLE_TENSORTYPES(DECL_INTERACTIVE_ACTIONS_TENSOR_OVERRIDE,
-                                   TensorSliced, T)
-
-    Tensor<T> contiguous() const;
-
-    TensorSliced<T> slice_unsqueeze(int i);
-    TensorSliced<T> const_slice_unsqueeze(int i) const;
-    TensorSliced<T> slice_squeeze(int i);
-    TensorSliced<T> const_slice_squeeze(int i) const;
-    TensorSliced<T> slice_squeeze(vector<int> dims);
-    TensorSliced<T> const_slice_squeeze(vector<int> dims) const;
-};
-
-template <typename T>
-class TensorTransposed : public Tensor<T> {
-    friend class Tensor<T>;
-    const shape_t old_strides;
-
-   public:
-    ~TensorTransposed() override = default;  // doesn't delete data
-    TensorTransposed(const Tensor<T>& t, const shape_t& permute_indexes) : 
-        Tensor<T>::Tensor(t.data, t.shape), old_strides(t.strides)
-    {
-        size_t new_shape(t.dim());
-        size_t new_strides(t.dim());
-        for (int i = 0; i < t.dim(); ++i){
-            size_t new_i = permute_indexes[i];
-            strides[i] = old_strides[new_i];
-            shape[i] = t.shape[new_i];
-        }
-    }
-    TensorTransposed(Tensor<T>& t, const shape_t& permute_indexes): 
-        Tensor<T>::Tensor(t.data, t.shape), old_strides(t.strides)
-    {
-        size_t new_shape(t.dim());
-        size_t new_strides(t.dim());
-        for (int i = 0; i < t.dim(); ++i){
-            size_t new_i = permute_indexes[i];
-            strides[i] = old_strides[new_i];
-            shape[i] = t.shape[new_i];
-        }
-    }
-
-    template <typename T_>
-    struct elem_iterator {
-        T_* data;
-        SliceGroup::const_iterator sg_iterator;
-        const shape_t shape;
-        const shape_t strides;
-
-        using difference_type = ptrdiff_t;
-        using value_type = T_;
-        using reference = T_&;
-        using pointer = T_*;
-        using iterator_category = std::forward_iterator_tag
-        
-        inline reference operator*() {
-            const auto& idx = *sg_iterator;
-            auto true_idx = ravel_index(idx, shape, strides);
-            return data[true_idx];
-        }
-
-        inline elem_iterator& operator++() {
-            ++sg_iterator;
-            return *this;
-        }
-
-        inline elem_iterator operator++(int) {
-            elem_iterator temp(*this);
-            ++(*this);
-            return temp;
-        }
-
-        inline bool operator==(const elem_iterator& other) {
-            return sg_iterator == other.sg_iterator;
-        }
-
-        inline bool operator!=(const elem_iterator& other) {
-            return !(*this == other);
-        }
-    };
-    using eiterator = elem_iterator<T>;
-    using ceiterator = elem_iterator<const T>;
-    inline eiterator elem_begin(TensorSliced& ts) {
-        
-    }
-    static inline eiterator elem_end(TensorSliced& ts);
-    static inline ceiterator const_elem_begin(const TensorSliced& ts);
-    static inline ceiterator const_elem_end(const TensorSliced& ts);
-};
-
-SliceGroup broadcast_index(const index_t& src_idx, const shape_t& src_shape,
-                           const shape_t& dst_shape);
-
-/**
- * Broadcasts an index of element in tensor towards an output shape, and
- * retrieves the broadcasted second source.
- * @param src1_idx the broadcasted index.
- * @param src1_shape the shape of the first input.
- * @param src2_shape the shape of the other input.
- * @param dst_shape the shape of output.
- * @return tuple of (bc_slicegroup_src2, bc_slicegroup_dst)
- */
-tuple<SliceGroup /*src2*/, SliceGroup /*dst*/> broadcast_index(
-    index_t src1_idx, const shape_t& src1_shape, const shape_t& src2_shape,
-    const shape_t& dst_shape);
-
-template <>
-class Tensor<bool> {};
-
-template <typename T>
-class TensorSparse : public Tensor<T> {
-    friend class Tensor<T>;
-    // TODO - implement.
-};
-
 using DoubleTensor = Tensor<double>;
-
-/**
- * Broadcasts shapes to a new output shape.
- * @param s1 first input shape
- * @param s2 second input shape
- * @param safe indicator of whether or not to validate shapes.
- * @return output shape.
- */
-inline shape_t broadcast_shapes(const shape_t& s1, const shape_t& s2,
-                                bool safe = false) {
-    if (s1.empty() || s2.empty()) return s1.empty() ? s2 : s1;
-    shape_t result(std::max(s1.size(), s2.size()));
-    for (int i = 0; i < result.size(); ++i) {
-        size_t e1 = i >= s1.size() ? 1 : s1[s1.size() - 1 - i];
-        size_t e2 = i >= s2.size() ? 1 : s2[s2.size() - 1 - i];
-        if (!safe && (e1 != e2 && e1 != 1 && e2 != 1))
-            throw broadcast_failure(s1, s2);
-        result[result.size() - 1 - i] = std::max(e1, e2);
-    }
-    return result;
-}
-
 }  // namespace blas
 
-#endif  // TARGETPRACTICE_TENSOR_H
+ 
+
+#endif // BLAS_TENSOR_H_
