+++ blas/implementation.cpp
@@ -3,6 +3,9 @@
 //
 
 #include "Tensor.h"
+#include "TensorView.h"
+#include "TensorSliced.h"
+#include "TensorTransposed.h"
 
 #include <algorithm>
 #include <iomanip>
